twitterdrunk/requirements.txt
argparse==1.2.1
oauthlib==0.6.3
python-dateutil==2.2
requests==2.1.0
requests-oauthlib==0.4.0
simplejson==3.6.4
six==1.8.0
twython==3.1.2
wsgiref==0.1.2
